+++ src/backend/commons/common/build.gradle
@@ -30,6 +30,7 @@ dependencies {
     api 'com.fasterxml.jackson.core:jackson-annotations'
     api 'org.apache.commons:commons-lang3'
     api 'org.apache.commons:commons-collections4'
+    api 'org.springframework.boot:spring-boot'
     implementation 'io.springfox:springfox-swagger2'
     implementation 'net.sf.dozer:dozer'
     implementation 'net.sourceforge.jchardet:jchardet'
