+++ src/backend/job-analysis/service-job-analysis/src/main/java/com/tencent/bk/job/analysis/service/remote/ApplicationServiceImpl.java
@@ -26,7 +26,7 @@ package com.tencent.bk.job.analysis.service.remote;
 
 import com.tencent.bk.job.analysis.client.ApplicationResourceClient;
 import com.tencent.bk.job.analysis.service.ApplicationService;
-import com.tencent.bk.job.manage.model.inner.ServiceApplicationDTO;
+import com.tencent.bk.job.common.model.ServiceApplicationDTO;
 import lombok.extern.slf4j.Slf4j;
 import lombok.val;
 import org.springframework.beans.factory.annotation.Autowired;
