+++ src/backend/job-analysis/service-job-analysis/src/main/java/com/tencent/bk/job/analysis/util/calc/AppMomYoyCalculator.java
@@ -27,7 +27,7 @@ package com.tencent.bk.job.analysis.util.calc;
 import com.fasterxml.jackson.core.type.TypeReference;
 import com.tencent.bk.job.common.statistics.model.dto.StatisticsDTO;
 import com.tencent.bk.job.common.util.json.JsonUtils;
-import com.tencent.bk.job.manage.model.inner.ServiceApplicationDTO;
+import com.tencent.bk.job.common.model.ServiceApplicationDTO;
 
 import java.util.List;
