+++ src/backend/job-manage/service-job-manage/src/main/java/com/tencent/bk/job/manage/api/web/impl/WebGlobalSettingsQueryResourceImpl.java
@@ -25,6 +25,7 @@
 package com.tencent.bk.job.manage.api.web.impl;
 
 import com.tencent.bk.job.analysis.consts.AnalysisConsts;
+import com.tencent.bk.job.common.constant.ErrorCode;
 import com.tencent.bk.job.common.constant.JobConstants;
 import com.tencent.bk.job.common.constant.ResourceScopeTypeEnum;
 import com.tencent.bk.job.common.iam.constant.ActionId;
@@ -33,6 +34,7 @@ import com.tencent.bk.job.common.iam.service.AppAuthService;
 import com.tencent.bk.job.common.model.Response;
 import com.tencent.bk.job.common.model.dto.AppResourceScope;
 import com.tencent.bk.job.common.model.dto.ApplicationDTO;
+import com.tencent.bk.job.common.util.feature.FeatureToggle;
 import com.tencent.bk.job.manage.api.web.WebGlobalSettingsQueryResource;
 import com.tencent.bk.job.manage.auth.NoResourceScopeAuthService;
 import com.tencent.bk.job.manage.config.JobManageConfig;
@@ -213,9 +215,13 @@ public class WebGlobalSettingsQueryResourceImpl implements WebGlobalSettingsQuer
         AppResourceScope appResourceScope = new AppResourceScope(scopeType, scopeId, null);
         ApplicationDTO applicationDTO = applicationService.getAppByScope(appResourceScope);
         if (applicationDTO != null) {
-            return Response.buildSuccessResp(appAuthService.getBusinessApplyUrl(appResourceScope));
-            // TODO:灰度开启
-            // return Response.buildCommonFailResp(ErrorCode.NEED_APP_SET_CONFIG);
+            // 普通业务/开启了CMDB业务集特性的业务集：从权限中心接口获取权限申请URL链接
+            if (appResourceScope.getType() == ResourceScopeTypeEnum.BIZ
+                || FeatureToggle.isCmdbBizSetEnabledForApp(appResourceScope.getAppId())) {
+                return Response.buildSuccessResp(appAuthService.getBusinessApplyUrl(appResourceScope));
+            }
+            // 未开启CMDB业务集特性的业务集：需要联系Job管理员添加为运维人员
+            return Response.buildCommonFailResp(ErrorCode.NEED_APP_SET_CONFIG);
         } else {
             return Response.buildSuccessResp(appAuthService.getBusinessApplyUrl(null));
         }
